docs: add Apache License version 2.0 to the project
The Apache License version 2.0 has been added to the project. This license grants users the permissions to use, reproduce, and distribute the work, as well as the ability to create derivative works. Additionally, it grants a patent license to users for any necessary patent claims. This addition ensures that the project is properly licensed and provides clarity on the permissions granted to users.